Fix incorrect port mapping in docker-compose
Fixed an incorrect port mapping the the docker-compose example as well
as added a restart attribute, since it's most likely a good use case to
always keep the server up and running, even if it fails or the host is
updated/restarted.

Also added a missing end parenthesis in the table of env variables.