Added LDAP authentication
If the builtin authentication fails, checks the username and password
against a LDAP server

Allow to create user without password since we don't store the user
password inside the database.

Every function that has the middleware.authengine
decorator will be used as authentication engine.
The order of the engines is based on the value of sequence,
lower sequences go first